[refactor/#155] 데이트포매터 리팩토링
